eeimg: Add Low-Swing app parameter
In order to acticate the low-swing mode (LSG1=0x3, LSG2=0x4) a parameter
'-l' must be passed to the program (by deault it's disabled). In this
case the corresponding initialization pattern is dumped in a cncbp
image. This mode shall reduce the power consumption dramatically so
to cool the IDT PCIe chip down a bit.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>