eeimg: cncbp: Add DS port app arguments
In order to make this firmware generator a but more generic,
we added a functionality, which lets to a change the IDT89HPESx
ports mode so make being downstream. It is necessary for our new
peripheral boards, which a based on ordinary endpoints - multi-IO
cards based on Asix AX99100 chips.

In ordet to change the ports mode, a user needes to update the IDT
PCIe switch firmware with one generated by this app, which argument
would be "-d 4,6,8,12,16,20" or any subset of this ports list.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>